+++ inject/inject-server/src/main/scala/com/twitter/inject/server/TwitterServer.scala
@@ -2,14 +2,12 @@ package com.twitter.inject.server
 
 import com.google.inject.Module
 import com.twitter.finagle.client.ClientRegistry
-import com.twitter.finagle.http.HttpMuxer
 import com.twitter.inject.Logging
 import com.twitter.inject.app.App
 import com.twitter.inject.modules.StatsReceiverModule
 import com.twitter.inject.utils.Handler
 import com.twitter.server.Lifecycle.Warmup
-import com.twitter.server.handler.ReplyHandler
-import com.twitter.server.internal.{FinagleBuildRevision, PromoteToOldGenUtils}
+import com.twitter.server.internal.FinagleBuildRevision
 import com.twitter.util.Await
 import com.twitter.util.registry.Library
 
@@ -95,13 +93,11 @@ trait TwitterServer
     Library.register(libraryName, Map[String, String]())
   }
 
-  /**
-   * After warmup completes, we want to run PromoteToOldGen without also signaling
-   * that we're healthy since we haven't successfully started our servers yet
-   */
   override protected def beforePostWarmup() {
     super.beforePostWarmup()
-    PromoteToOldGenUtils.beforeServing()
+
+    // trigger gc before accepting traffic
+    prebindWarmup()
   }
 
   override protected def postWarmup() {
@@ -122,6 +118,6 @@ trait TwitterServer
   override protected def afterPostWarmup() {
     super.afterPostWarmup()
     info("Enabling health endpoint on port " + PortUtils.getPort(adminHttpServer))
-    HttpMuxer.addHandler("/health", new ReplyHandler("OK\n"))
+    warmupComplete()
   }
 }